Fold UdpState into UdpSocketState
Simplifies the quinn-udp API, and helps isolate internal complexity to
the platforms that need it.
